feat: Add elim constraint elaboration
Whenever an elimination constraint is not satisfied, then it adds the constraint
as a new local one for the definition. Currently behind a flag to avoid
backward compatibility for the moment.

fix: Reintroduce QElimTo to univProblem

It was replaced by QLeq to deal only with cumulativity
constraints.

refactor: Remove unnecessary abstraction of code

Maybe it's using a more general API than it should, considering
PConstraints instead of simply UnivConstrants.