feat: Remove elim constraint Type->s by default
fix: Update to have dominance check not be done on same quality

Notably, dominance updates between sort variables were not checking
if the variables were the same. If both were the same and they were
not dominated (e.g. when removing the default Type->s constraint),
then it would loop because it would postpone the dominance check.
It really doesn't make sense though to update the dominance of a
sort variable when adding the reflexive elimination constraint.
The dominant will be the same in both cases, and if there is none,
then it will be dominated eventually when checking with other sorts.